+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 #### Version 3.8.1 (23.02.2024)
 - **Fix**: Wrong "O" removed from key cap
+- **Fix**: Keyboard heatmap precision has been fixed (it now shows 99.838% and 8.7%)
 
 #### Version 3.8 (23.02.2024)
 - **New**: Airbnb eslint rules added